Slår sammen testfiler for nytt arbeidsforhold i nytt_arbeidsforhold.spec.ts
